Move MCP histogram test out of conflicting file (#3782)
