+++ tests/CodeIndex.Tests/McpServerTests.cs
@@ -111,31 +111,6 @@ public class McpServerTests : IDisposable
         Assert.False(MethodCallsType(method!, typeof(SpinWait)));
     }
 
-    [Fact]
-    public void BuildTopFileHistogram_OrdersCountsDeterministically_Issue3782()
-    {
-        var method = typeof(McpServer).GetMethod("BuildTopFileHistogram", BindingFlags.NonPublic | BindingFlags.Instance);
-        Assert.NotNull(method);
-        var generic = method!.MakeGenericMethod(typeof(SearchResult));
-        var results = new[]
-        {
-            new SearchResult { Path = "src/b.cs" },
-            new SearchResult { Path = "src/a.cs" },
-            new SearchResult { Path = "src/b.cs" },
-            new SearchResult { Path = "src/c.cs" },
-            new SearchResult { Path = "src/c.cs" },
-        };
-
-        var histogram = Assert.IsType<JsonArray>(generic.Invoke(_server, [results, new Func<SearchResult, string?>(result => result.Path)]));
-
-        Assert.Equal("src/b.cs", histogram[0]!["path"]!.GetValue<string>());
-        Assert.Equal(2, histogram[0]!["count"]!.GetValue<int>());
-        Assert.Equal("src/c.cs", histogram[1]!["path"]!.GetValue<string>());
-        Assert.Equal(2, histogram[1]!["count"]!.GetValue<int>());
-        Assert.Equal("src/a.cs", histogram[2]!["path"]!.GetValue<string>());
-        Assert.Equal(1, histogram[2]!["count"]!.GetValue<int>());
-    }
-
     private static bool MethodCallsType(MethodInfo method, Type declaringType)
     {
         var body = method.GetMethodBody()?.GetILAsByteArray();
